requirements.txt
anthropic>=0.39.0
pyautogui>=0.9.54
Pillow>=10.0.0
pyobjc-framework-Quartz
python-dotenv>=1.0.0
